functions/requirements.txt
functions-framework
google-cloud-storage
google-genai
